Move widgets back to main QT deps line
Widgets is not the SAME AS QTwidgets and does not belong with QTWIDGETS.

In mythtv the qt widget is  a part of the main gui design. QTWIDGETS are specific to QT. Sure but they both just called widgets but they dont belong together as deps. This obscurity will eventually cause problems when webkit is finally cut out of mythtv upstream and replaced with new classes. Even then , when qtwidgets are no longer needed, widgets will still be needed as part of the main gui.